+++ CHANGELOG.md
@@ -38,6 +38,8 @@
   - [#4482](https://github.com/wix/ricos/pull/4482) fix external formatting toolbar buttons obj name
 - `divider`
   - [#4483](https://github.com/wix/ricos/pull/4483) fix defaults on tiptap
+- `divider`
+  - [#4487](https://github.com/wix/ricos/pull/4487) fix disabled button & fullWidth when aligned
 
 ## 8.72.36 (Aug 24, 2022)